+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/LogWatchCallback.java
@@ -18,7 +18,6 @@ package io.fabric8.kubernetes.client.dsl.internal;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.dsl.LogWatch;
-import io.fabric8.kubernetes.client.utils.BlockingInputStreamPumper;
 import io.fabric8.kubernetes.client.utils.InputStreamPumper;
 import io.fabric8.kubernetes.client.utils.Utils;
 import okhttp3.Call;
@@ -42,7 +41,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import static io.fabric8.kubernetes.client.utils.Utils.closeQuietly;
-import static io.fabric8.kubernetes.client.utils.Utils.shutdownExecutorService;
 
 public class LogWatchCallback implements LogWatch, Callback, AutoCloseable {
 
@@ -56,8 +54,6 @@ public class LogWatchCallback implements LogWatch, Callback, AutoCloseable {
     private final CompletableFuture<Void> startedFuture = new CompletableFuture<>();
     private final ExecutorService executorService = Executors.newSingleThreadExecutor();
     private final AtomicBoolean closed = new AtomicBoolean(false);
-
-    private InputStreamPumper pumper;
 
     @Deprecated
     public LogWatchCallback(OutputStream out) {
@@ -95,23 +91,19 @@ public class LogWatchCallback implements LogWatch, Callback, AutoCloseable {
 
     /**
      * Performs the cleanup tasks:
-     * 1. closes the InputStream pumper
+     * 1. cancels the InputStream pumper
      * 2. closes all internally managed closeables (piped streams).
      *
      * The order of these tasks can't change or its likely that the pumper will through errors,
      * if the stream it uses closes before the pumper it self.
      */
     private void cleanUp() {
-      try {
-        if (!closed.compareAndSet(false, true)) {
-          return;
-        }
-
-        closeQuietly(pumper);
-        shutdownExecutorService(executorService);
-      } finally {
-        closeQuietly(toClose);
+      if (!closed.compareAndSet(false, true)) {
+        return;
       }
+
+      executorService.shutdownNow();
+      closeQuietly(toClose);
     }
 
     public void waitUntilReady() {
@@ -140,19 +132,12 @@ public class LogWatchCallback implements LogWatch, Callback, AutoCloseable {
 
     @Override
     public void onResponse(Call call, final Response response) throws IOException {
-       pumper = new BlockingInputStreamPumper(response.body().byteStream(), input -> {
-           try {
-               out.write(input);
-           } catch (IOException e) {
-               throw KubernetesClientException.launderThrowable(e);
-           }
-       }, () -> {
-         cleanUp();
-         response.close();
-       });
-
       if (!executorService.isShutdown()) {
-        executorService.submit(pumper);
+        // the task will be cancelled via shutdownNow
+        InputStreamPumper.pump(response.body().byteStream(), out::write, executorService).whenComplete((o, t) -> {
+          cleanUp();
+          response.close();
+        });
         startedFuture.complete(null);
       }